requirements.txt
asgiref==3.7.2
Django==4.2.6
sqlparse==0.4.4
typing_extensions==4.8.0
psycopg2==2.9.3
djangorestframework == 3.14
markdown
django-filter
djangorestframework_simplejwt
python-dotenv~=1.0.0
django-rosetta
django_jalali
django-redis
Pillow
celery~=5.3.5
django-celery-results
django_celery_beat
drf_spectacular
django-role-permissions
factory_boy
django-guardian